requirements.txt
Flask==0.10
Pillow>=2.0.0
coverage==3.5
html5lib==0.90
httplib2==0.7.6
nose==1.0.0
pyPdf==1.13
reportlab==2.5
flake8==2.0
lob-python==1.0